start(app): refactoring
