+++ src/assets/icon/app/burger-menu/index.ts
@@ -1,5 +1,3 @@
-import { Partners } from './partners'
-import { Reasons } from './reasons'
-import { Reviews } from './reviews'
-
-export { Partners, Reviews, Reasons }
+export * from './partners'
+export * from './reasons'
+export * from './reviews'
