+++ src/assets/icon/app/index.ts
@@ -1,13 +1,4 @@
-import { FullStar } from './fullStar'
-import { HalfEmptyStar } from './halfEmptyStar'
-import { LogoIcon } from './logo'
-import { Partners, Reviews, Reasons } from './burger-menu'
-
-export {
-  FullStar,
-  HalfEmptyStar,
-  LogoIcon,
-  Partners,
-  Reviews,
-  Reasons,
-}
+export * from './burger-menu/index'
+export * from './fullStar'
+export * from './logo'
+export * from './halfEmptyStar'
