+++ src/assets/icon/border/index.ts
@@ -1,15 +1,6 @@
-import { AddBoard } from './AddBoard'
-import { Filter } from './Filter'
-import { Menu } from './Menu'
-import { Search } from './Search'
-import { SecondMenu } from './SecondMenu'
-import { Sort } from './Sort'
-
-export {
-  AddBoard,
-  Filter,
-  Menu,
-  Sort,
-  Search,
-  SecondMenu,
-}
+export * from './Filter'
+export * from './Menu'
+export * from './AddBoard'
+export * from './Sort'
+export * from './Search'
+export * from './SecondMenu'
