+++ src/assets/icon/index.ts
@@ -1,34 +1,5 @@
-// APP
-import {
-  FullStar,
-  Reasons,
-  Reviews,
-  Partners,
-  LogoIcon,
-  HalfEmptyStar,
-} from './app'
+/* app */
+export * from './app'
 
-// BOARD
-import {
-  AddBoard,
-  SecondMenu,
-  Search,
-  Sort,
-  Menu,
-  Filter,
-} from './border'
-
-export {
-  FullStar,
-  Reviews,
-  Reasons,
-  Partners,
-  LogoIcon,
-  HalfEmptyStar,
-  AddBoard,
-  Sort,
-  Search,
-  SecondMenu,
-  Menu,
-  Filter,
-}
+/* border */
+export * from './border'
