+++ src/board/layouts/Main/components/index.ts
@@ -1,5 +1,3 @@
-import { BoardItems } from './BoardItems'
-import { BoardList } from './BoardList'
-import { SubHeader } from './SubHeader'
-
-export { BoardList, BoardItems, SubHeader }
+export * from './SubHeader'
+export * from './BoardList'
+export * from './BoardItems'
