+++ src/core/components/app/index.ts
@@ -1,5 +1,3 @@
-import { Anchor } from './Anchor'
-import { Component } from './Component'
-import { Layout } from './Layout'
-
-export { Anchor, Component, Layout }
+export * from './Layout'
+export * from './Anchor'
+export * from './Component'
