+++ src/core/components/common/Header/components/UserWrapper/index.tsx
@@ -1,15 +1,15 @@
-import React, { useContext } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { observer } from 'mobx-react'
 import { useLocation, useNavigate } from 'react-router-dom'
 
 import { Button } from '../../../../ui-components'
-import { AppContext } from '../../../../../../App'
+import { useUser } from '../../../../../../domain/hooks/useUser'
 
 export const UserWrapper = observer(() => {
   const navigate = useNavigate()
   const { pathname } = useLocation()
-  const { userStore } = useContext(AppContext)
+  const { user } = useUser()
   const { t } = useTranslation()
 
   const onUserLogin = () => {
@@ -20,8 +20,8 @@ export const UserWrapper = observer(() => {
   return (
     <div>
       {
-        userStore.profile
-          ? <h3>{userStore.profile.email}</h3>
+        user.profile
+          ? <h3>{user.profile.email}</h3>
           : <Button description={t(`translation:user-login`)} onClick={onUserLogin} />
       }
     </div>
