+++ src/core/components/index.ts
@@ -1,38 +1,8 @@
-/* APP */
-import { Anchor, Layout, Component } from './app'
+/* app */
+export * from './app'
 
-/* COMMON */
-import { Header, NotFound, Loader } from './common'
+/* common */
+export * from './common'
 
-/* UI */
-import {
-  Button,
-  Checkbox,
-  FormGroup,
-  Hint,
-  Icon,
-  Image,
-  Input,
-  Link,
-  Logo,
-  Select,
-} from './ui-components'
-
-export {
-  Anchor,
-  Layout,
-  Component,
-  Header,
-  NotFound,
-  Loader,
-  Button,
-  Checkbox,
-  FormGroup,
-  Hint,
-  Icon,
-  Image,
-  Input,
-  Link,
-  Logo,
-  Select,
-}
+/* ui */
+export * from './ui-components'
