+++ src/core/components/ui-components/index.ts
@@ -1,23 +1,10 @@
-import { FormGroup } from './FormGroup'
-import { Checkbox } from './Checkbox'
-import { Button } from './Button'
-import { Select } from './Select'
-import { Image } from './Image'
-import { Input } from './Input'
-import { Hint } from './Hint'
-import { Link } from './Link'
-import { Icon } from './Icon'
-import { Logo } from './Logo'
-
-export {
-  FormGroup,
-  Checkbox,
-  Select,
-  Button,
-  Image,
-  Input,
-  Hint,
-  Link,
-  Icon,
-  Logo,
-}
+export * from './Icon'
+export * from './Logo'
+export * from './Hint'
+export * from './Button'
+export * from './Link'
+export * from './Checkbox'
+export * from './FormGroup'
+export * from './Image'
+export * from './Input'
+export * from './Select'
