+++ src/core/hooks/index.ts
@@ -1,17 +1,7 @@
-import { useFetch } from './useFetch/useFetch'
-import { usePosition } from './usePosition'
-import { useScroll } from './useScroll'
-import { useSlots } from './useSlots'
-import { useDispatch } from './useDispatch'
-import { useMatchMedia } from './useMatchMedia'
-import { useClickOutside } from './useClickOutside'
-
-export {
-  useClickOutside,
-  useMatchMedia,
-  usePosition,
-  useDispatch,
-  useScroll,
-  useSlots,
-  useFetch,
-}
+// export * from './useFetch'
+export * from './useScroll'
+export * from './useDispatch'
+export * from './usePosition'
+export * from './useSlots'
+export * from './useClickOutside'
+export * from './useMatchMedia'
