+++ src/core/lib/utilities/index.ts
@@ -1,0 +1,2 @@
+export * from './functions'
+export * from './storage'
