+++ src/domain/hooks/useUser/interface/index.ts
@@ -1,0 +1,5 @@
+import { User } from '../../../user'
+
+export interface IUseUserResult {
+  user: User
+}
