+++ src/public/layouts/Main/components/PublicAssessment/components/index.ts
@@ -1,0 +1,1 @@
+export * from './Score'
