+++ src/public/layouts/Main/components/index.ts
@@ -1,13 +1,5 @@
-import { PublicReviews } from './PublicReviews'
-import { PublicHelp } from './PublicHelp'
-import { PublicAssessment } from './PublicAssessment'
-import { PublicReasons } from './PublicReasons'
-import { PublicPartners } from './PublicPartners'
-
-export {
-  PublicReviews,
-  PublicHelp,
-  PublicAssessment,
-  PublicReasons,
-  PublicPartners,
-}
+export * from './PublicAssessment'
+export * from './PublicHelp'
+export * from './PublicReasons'
+export * from './PublicPartners'
+export * from './PublicReviews'
