+++ src/public/modules/login/components/index.ts
@@ -1,4 +1,2 @@
-import { AuthWrapper } from './AuthWrapper'
-import { LogoList } from './LogoList'
-
-export { AuthWrapper, LogoList }
+export * from './LogoList'
+export * from './AuthWrapper'
