+++ src/public/modules/login/index.ts
@@ -1,0 +1,2 @@
+export * from './components'
+export * from './views'
