+++ src/public/modules/login/views/Login/components/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginForm'
