+++ src/public/modules/login/views/Login/hooks/useLogin/index.ts
@@ -1,23 +1,20 @@
-import { useContext } from 'react'
 import jwtDecode from 'jwt-decode'
 import { useNavigate } from 'react-router-dom'
-import { Storage } from '../../../../../../../core/lib/utilities/storage'
-import { IUser } from '../../../../../../../domain/user/interface'
+import { Storage, useDispatch } from '../../../../../../../core'
+import { IUser, loginUser } from '../../../../../../../domain'
 import { ILoginForm } from '../../components/LoginForm/interface'
-import { useDispatch } from '../../../../../../../core/hooks'
-import { loginUser } from '../../../../../../../domain/user/service'
 import { IUseLoginReturn } from './interface'
-import { AppContext } from '../../../../../../../App'
+import { useUser } from '../../../../../../../domain/hooks/useUser'
 
 export const useLogin = (): IUseLoginReturn => {
-  const { userStore } = useContext(AppContext)
+  const { user } = useUser()
   const navigate = useNavigate()
   const { dispatch, data: response } = useDispatch(loginUser)
 
   const updateUserInformation = () => {
     const storage = new Storage('local')
-    const user: IUser = jwtDecode(storage.getItem('jwt-token'))
-    userStore.setUser(user)
+    const userProfile: IUser = jwtDecode(storage.getItem('jwt-token'))
+    user.setUser(userProfile)
   }
 
   const onLogin = async (data: ILoginForm) => {
