+++ src/public/modules/login/views/Login/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { AuthWrapper } from '../../components'
-import { LoginForm } from './components/LoginForm'
+import { LoginForm } from './components'
 
 export const LoginPage = () => {
   return (
