+++ src/public/modules/login/views/Registration/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useRegistration'
