+++ src/public/modules/login/views/Registration/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { AuthWrapper } from '../../components'
-import { RegistrationForm } from './components/RegistrationForm'
+import { RegistrationForm } from './components'
 
 export const RegistrationPage = () => {
   return (
