+++ src/public/modules/login/views/index.ts
@@ -1,4 +1,2 @@
-import { LoginPage } from './Login'
-import { RegistrationPage } from './Registration'
-
-export { LoginPage, RegistrationPage }
+export * from './Login'
+export * from './Registration'
